requirements.txt
# ---- Core dependencies ----
python-dotenv==1.0.1
supabase==2.5.1
requests==2.32.3
websockets==12.0
kucoin-python==1.0.26

# ---- Logging + Utilities ----
urllib3==2.2.3
certifi==2025.6.15
charset-normalizer==3.4.0
idna==3.10
